scripts/requirements.txt
torch
pandas
tqdm
datasets
sentence-splitter
transformers
nltk
accelerate
pytest
python-dotenv
bitsandbytes
pydantic
langchain
stanza
